+++ package.json
@@ -9,7 +9,7 @@
     "cypress:interactive": "cypress open",
     "cypress:storybook": "cypress run --project ./.storybook/.",
     "cypress:storybook:interactive": "cypress open --project ./.storybook/.",
-    "data-validate": "node src/app/datavalidator.jsx",
+    "data-validate": "node src/app/datavalidator/datavalidator.jsx",
     "dev": "razzle start",
     "lighthouse": "lighthouse http://localhost:7080/article/id --output-path=./reports/report.html --view",
     "postmerge": "./scripts/dependencyUpdateCheck.sh",
